+++ include/Sensors/aht20.hpp
@@ -7,7 +7,7 @@
 #ifdef DEVICE_AHT20
 
 // Function declarations
-uint32_t init_aht20();
+uint8_t init_aht20();
 void read_aht20(transmit_data_t *temp, transmit_data_t *humidity);
 
 #endif // DEVICE_AHT20
