+++ include/Sensors/capacitive_soil_moisture.hpp
@@ -7,7 +7,7 @@
 #ifdef DEVICE_CAPACITIVE_SOIL_MOISTURE_SENSOR
 
 // Function declarations
-uint32_t init_soil_sensor();
+uint8_t init_soil_sensor();
 void read_capacitive_soil_moisture_sensor(transmit_data_t *moisture);
 
 #endif // DEVICE_CAPACITIVE_SOIL_MOISTURE_SENSOR
