+++ include/Sensors/ens160.hpp
@@ -7,7 +7,7 @@
 #ifdef DEVICE_ENS160
 
 // Function declarations
-uint32_t init_ens160();
+uint8_t init_ens160();
 void read_ens160(transmit_data_t *aqi, transmit_data_t *tvoc, transmit_data_t *eCO2, transmit_data_t *temp, transmit_data_t *humidity);
 
 #endif // DEVICE_ENS160
